requirements.txt
asgiref==3.6.0
autopep8==2.0.2
dj-database-url==1.3.0
Django==4.2
django-autoslug==1.9.9
django-etc==1.4.0
Pillow==9.5.0
psycopg2==2.9.6
pycodestyle==2.10.0
python-decouple==3.8
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
whitenoise==6.4.0
